wip: scalajs support
