+++ core/js/src/main/scala/io/github/timwspence/cats/stm/STMConstants.scala
@@ -1,0 +1,29 @@
+/*
+ * Copyright 2017-2021 TimWSpence
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.github.timwspence.cats.stm
+
+object STMConstants {
+  final val PureT        = 0
+  final val AllocT       = 1
+  final val BindT        = 2
+  final val HandleErrorT = 3
+  final val GetT         = 4
+  final val ModifyT      = 5
+  final val OrElseT      = 6
+  final val AbortT       = 7
+  final val RetryT       = 8
+}
