+++ core/jvm/src/main/java/io/github/timwspence/cats/stm/STMConstants.java
