+++ core/jvm/src/test/scala/io/github/timwspence/cats/stm/PropertyTests.scala
