+++ core/shared/src/main/scala/io/github/timwspence/cats/stm/STMLike.scala
